CI: Better caching, minor improvements
The documentation for the `haskell/actions/setup` action demonstrated a
nice improvement to caching, where a cache is created even if the
project itself fails to build. It also demonstrated cache invalidation
based on hashes of all the dependencies, which improves caching
efficiency.

Also pass `--pedantic` to `stack build` and don't invoke
`cabal v2-update` unnecesarily (it's already done by the setup action)