CI: Simplify Stack version bumps
The `.ci` directory contains `stack.yaml` files for each GHC version we
test with Stack on CI. These files used to be named after the full GHC
version, which is annoying if you want to bump to a later minor version.
The files are now named only after the major GHC version so the file
doesn't change name on such a bump.