Tighten lower bounds, loosen upper bounds
We now explicitly test in CI the lower and upper bounds in our
`build-depends`.

The PvP allows incrementing just the minor version of a package in the
case of an API addition. Testing the lower bounds explicitly assures we
don't accidentally rely on an API addition made in a minor version.

Attempting to test this case revealed our existing lower bounds could
not be tested because it did not lead to a valid build plan. So the
bounds have been tightened to specify a minor version that we can
actually test.

We can loosen our upper bounds to include the newest versions of `text`
and `tasty` by testing this configuration with cabal's `allow-newer`.